+++ app/src/main/java/app/gyrolet/mpvrx/preferences/SubtitlesPreferences.kt
@@ -41,9 +41,8 @@ class SubtitlesPreferences(
 
   val pickerPath = preferenceStore.getString("sub_picker_path")
   
-  val subdlApiKey = preferenceStore.getString("subdl_api_key", "")
   val subtitleSaveFolder = preferenceStore.getString("sub_save_folder", "")
-  val subdlLanguages = preferenceStore.getStringSet("subdl_languages", setOf("en"))
+  val subtitleSearchLanguages = preferenceStore.getStringSet("subdl_languages", setOf("en"))
   
   val wyzieSources = preferenceStore.getStringSet("wyzie_sources", setOf("all"))
   val wyzieFormats = preferenceStore.getStringSet("wyzie_formats", setOf("srt", "ass"))